+++ packages/app/src/flame/variations/parametric/general/vogelVar.tsx
@@ -45,7 +45,7 @@ export const vogelVar = parametricVariation(
   (pos, _varInfo, P) => {
     'use gpu'
     const dist = sqrt(pos.x * pos.x + pos.y * pos.y)
-    const i = floor(P.n * dist * 0.5)
+    const i = f32(floor(P.n * dist * 0.5))
     const theta = i * GOLDEN_ANGLE
     const r = P.scale * sqrt(i)
     return vec2f(r * cos(theta), r * sin(theta))
