+++ packages/app/src/flame/variations/simple/general/apocarpetVar.ts
@@ -1,4 +1,4 @@
-import { vec2f } from 'typegpu/data'
+import { f32, vec2f } from 'typegpu/data'
 import { floor, sqrt } from 'typegpu/std'
 import { random } from '@/shaders/random'
 import { simpleVariation } from '../types'
@@ -12,7 +12,7 @@ export const apocarpetVar = simpleVariation(
     let y = pos.y
     const r = 1.0 / (1.0 + sqrt(2.0))
     const denom = pos.x * pos.x + pos.y * pos.y + 1.0e-10
-    const weight = floor(6.0 * random())
+    const weight = f32(floor(6.0 * random()))
 
     if (weight === 0.0) {
       x = (2.0 * pos.x * pos.y) / denom
